Lots of changes
Working image for hardware CPC2.1.
The hardware for CPC2.2 is on it's way from OSH Park, so there is come code for the new board version 2.2.
Note particularly the SDRAM module - this is a key module for the next phase of the project.
Take a look at https://intelligenttoasters.blog/2017/10/12/retro-cpc-dongle-part-29/