Added a byte cache and a memory arbiter.
This code is very specific to the memory device that I selected, such as the R-C delay and the write settle time.
If a different device is used, or a 3-clock R-C delay is used, then this function will still work with some minor modifications.
Note that the cache itself is Altera specific, using the 2-port SRAM